+++ internal/assert/assert.go
@@ -4,11 +4,32 @@
 
 package assert
 
-import "testing"
+import (
+	"sort"
+	"strings"
+	"testing"
+)
 
 // Comparable asserts two comparable values that they equal.
 func Comparable[T comparable](t *testing.T, name string, want T, got T) {
 	if want != got {
 		t.Errorf(`%s: want "%v", got "%v"`, name, want, got)
+	}
+}
+
+// ElementsMatch asserts that two slices of strings contain the same values,
+// with no regard to the ordering.
+func ElementsMatch[S ~[]string](t *testing.T, want, got S) {
+	wantClone := make(S, len(want))
+	copy(wantClone, want)
+	sort.Strings(wantClone)
+	gotClone := make(S, len(got))
+	copy(gotClone, got)
+	sort.Strings(gotClone)
+
+	wantStr := strings.Join(wantClone, ", ")
+	gotStr := strings.Join(gotClone, ", ")
+	if wantStr != gotStr {
+		t.Errorf("want %q, got %q", wantStr, gotStr)
 	}
 }